% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_annotation.R
\name{av_annotate_regions}
\alias{av_annotate_regions}
\title{Annotate Variants with Custom Regions (e.g., HARs)}
\usage{
av_annotate_regions(obj, regions, label = "in_region")
}
\arguments{
\item{obj}{An object of class \code{AlphaVarSet}.}

\item{regions}{Path to a BED file (character) or a \code{GRanges} object.}

\item{label}{Character string. Name of the new column (default: "in_region").}
}
\value{
An \code{AlphaVarSet} object with a new boolean column indicating overlap.
}
\description{
Checks for overlap between variants and a set of genomic regions provided
either as a BED file or a GRanges object. Adds a boolean column to the data.
}
