% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_import.R
\name{av_import_csv}
\alias{av_import_csv}
\title{Import AlphaGenome Results}
\usage{
av_import_csv(file, col_map = list(), sep = ",")
}
\arguments{
\item{file}{Path to the CSV file.}

\item{col_map}{A named list mapping internal names to CSV headers.
Defaults to: list(score = "quantile_score", modality = "output_type",
variant_id = "variant_id", gene = "gene_name").}

\item{sep}{Separator for the file (default ",").}
}
\value{
An \code{AlphaVarSet} object.
}
\description{
Reads a CSV file and creates an AlphaVarSet object.
It standardizes column names based on a provided mapping.
}
