% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_viz_distribution.R
\name{av_plot_ecdf}
\alias{av_plot_ecdf}
\title{Plot Cumulative Distribution (ECDF)}
\usage{
av_plot_ecdf(obj, abs_score = TRUE)
}
\arguments{
\item{obj}{An \code{AlphaVarSet} object.}

\item{abs_score}{Logical.}
}
\value{
A ggplot object.
}
\description{
Plot Cumulative Distribution (ECDF)
}
