% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_viz_genome.R
\name{av_plot_locus}
\alias{av_plot_locus}
\title{Plot Genomic Locus (Manhattan-style)}
\usage{
av_plot_locus(obj, gene_name, flank_bp = 0, color_by = "modality")
}
\arguments{
\item{obj}{An \code{AlphaVarSet} object.}

\item{gene_name}{The gene to plot (filters variants for this gene).}

\item{flank_bp}{Number of base pairs to extend the plotting window (default 500).}

\item{color_by}{Column to map to color (default "modality").}
}
\value{
A ggplot object.
}
\description{
Visualizes variants along genomic coordinates for a specific gene or region.
}
