% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_qc.R
\name{av_plot_qc_dist}
\alias{av_plot_qc_dist}
\title{Plot Score Distribution (QC)}
\usage{
av_plot_qc_dist(obj, type = "histogram")
}
\arguments{
\item{obj}{An object of class \code{AlphaVarSet}.}

\item{type}{Plot type: "histogram" or "density".}
}
\value{
A ggplot object.
}
\description{
Visualizes the distribution of raw scores to check for artifacts or skew.
}
