% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_viz_distribution.R
\name{av_plot_ranking}
\alias{av_plot_ranking}
\title{Plot Ranked Boxplots}
\usage{
av_plot_ranking(obj, abs_score = TRUE)
}
\arguments{
\item{obj}{An \code{AlphaVarSet} object.}

\item{abs_score}{Logical. Use absolute scores?}
}
\value{
A ggplot object.
}
\description{
Displays boxplots of scores, sorted by median value or Wilcoxon shift.
}
