% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_qc.R
\name{av_qc_summary}
\alias{av_qc_summary}
\title{Generate QC Summary}
\usage{
av_qc_summary(obj)
}
\arguments{
\item{obj}{An object of class \code{AlphaVarSet}.}
}
\value{
A tibble with summary statistics.
}
\description{
Provides a quick overview of the dataset dimensions and missing values.
}
