% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_shiny.R
\name{av_run_shiny}
\alias{av_run_shiny}
\title{Launch AlphaVaR Interactive Explorer}
\usage{
av_run_shiny()
}
\value{
This function does not return; it interrupts R execution to run the app.
}
\description{
Launches a local Shiny application to visualize and explore variant scores interactively.
Users can upload CSV files, view volcano plots, and filter results.
}
