% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{new_AlphaVarSet}
\alias{new_AlphaVarSet}
\title{Constructor for AlphaVarSet objects}
\usage{
new_AlphaVarSet(data, meta = list())
}
\arguments{
\item{data}{A tibble containing the variant data.}

\item{meta}{A list containing metadata (filename, date).}
}
\value{
An object of class AlphaVarSet.
}
\description{
Internal function to create the object. Use av_import_csv() instead.
}
\keyword{internal}
