% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_viz_genome.R
\name{parse_coords}
\alias{parse_coords}
\title{Parse Coordinate Strings}
\usage{
parse_coords(coord_vec)
}
\description{
Converts strings like "chr1:100-200" into a tibble with chrom, start, end.
}
\keyword{internal}
