% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{print.AlphaVarSet}
\alias{print.AlphaVarSet}
\title{Print method for AlphaVarSet}
\usage{
\method{print}{AlphaVarSet}(x, ...)
}
\arguments{
\item{x}{An AlphaVarSet object.}

\item{...}{Additional arguments.}
}
\description{
Gives a nice summary in the console when typing the object name.
}
