% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{validate_AlphaVarSet}
\alias{validate_AlphaVarSet}
\title{Validator for AlphaVarSet objects}
\usage{
validate_AlphaVarSet(x)
}
\arguments{
\item{x}{An AlphaVarSet object.}
}
\value{
The object if valid, error otherwise.
}
\description{
Checks if the object is structurally sound.
}
\keyword{internal}
